Add workaround for CSDL-to-JSON including future properties in errata versions
per Jeff Autor:
The issue is when we force an errata schema to be generated for some reason, the down-rev’ed errata schemas are picking up any enums added (because there’s no differentiation in the CSDL).  And therefore the doc generator is seeing things like “Cancelled” added to Task/JobState in “v1.0.5+” instead of “v1.3+”…

What I’m thinking is if we could have the doc generator only add properties/enums from minor versions (v1.x.0) and only use errata to get ‘corrections’ to existing property definitions that would avoid this issue.