Fix a bug wherein enums were dropped for some properties of common objects.
This affected definitions in schemas that were not versioned at the "top" object for the file, but
rather at individual properties or definitions, and that further had an "anyOf" specifying a definition
that went straight to an enum. More generally, this would probably affect any pointer to a definition that
did not specify an object with properties. Example: AddressOrigin in IPAddresses/IPv4Address.