Doc generator: Update the schema_data with properties found in an unversioned file.
This fixes a case in the 2018.3 schemas, where some enums were moved to unversioned files, and
the doc generator wasn't finding the definitions.

Previously we updated property_data, but not the schema_data used for traversal.
This also exposed that we were potentially overwriting data, so we've stopped doing that --
this may be wrong, as the previous arrangement addressed cases where the unversioned file
had a definition "overriding" those in the versioned files. But applying this to the schema_data
exposed that sometimes the unversioned file has an "anyOf" list for all the versioned  files...